refactor(app): extract useWorkspaceRouteState — the session route's data core
The final and largest step of the session-route decomposition: the route's
data + navigation core moves verbatim to shell/use-workspace-route-state.ts
(855 lines):

- URL-derived selection (params, navigateToWorkspaceSession) and the 12 core
  states (loading/client/baseUrl/token/workspaces/order/sessions/errors/
  connection-overrides/routeError/legacy-selection/retrying)
- refreshRouteState + background session loading (pending-created merge,
  startup retry, per-workspace in-flight guards)
- endpoint + opencode client resolution (selectedWorkspaceEndpoint,
  opencodeClient, not-found message derivation)
- mount/visibility/settings listeners, desktop local-server reconnect,
  remote connection checks, last-session restore, welcome redirect, and the
  route inspector slice

Two leaks became explicit inputs: onServerSettingsChanged (settings-version
bump) and onHostInfo (local server host info), both owned by the route's
openwork-server settings cluster. isTransientStartupError moves to
route-workspaces.ts (used by both files now).

session-route.tsx: 2,564 -> 1,866 lines — 3,328 at the start of the
decomposition, a 44% reduction with every concern now in a named module.

Tested:
- pnpm typecheck clean; bun test tests/ 58 pass; pnpm build clean
- pnpm test:e2e — 23/23 assertions ok (boots the real app + opencode and
  exercises exactly the machinery this commit moved: workspace listing,
  session creation, sidebar loading)
- live: vite serves the new module graph (200)
